+++ pyfemtet/opt/visualization/process_monitor/pages.py
@@ -10,11 +10,11 @@ import pandas as pd
 from dash import Output, Input, State, callback_context, no_update, ALL
 from dash.exceptions import PreventUpdate
 
-from pyfemtet.opt.visualization2.wrapped_components import dcc, dbc, html
-from pyfemtet.opt.visualization2.base import AbstractPage, logger
-from pyfemtet.opt.visualization2.complex_components.main_graph import MainGraph  # , FLEXBOX_STYLE_ALLOW_VERTICAL_FILL
-from pyfemtet.opt.visualization2.complex_components.control_femtet import FemtetControl, FemtetState
-from pyfemtet.opt.visualization2.complex_components.alert_region import AlertRegion
+from pyfemtet.opt.visualization.wrapped_components import dcc, dbc, html
+from pyfemtet.opt.visualization.base import AbstractPage, logger
+from pyfemtet.opt.visualization.complex_components.main_graph import MainGraph  # , FLEXBOX_STYLE_ALLOW_VERTICAL_FILL
+from pyfemtet.opt.visualization.complex_components.control_femtet import FemtetControl, FemtetState
+from pyfemtet.opt.visualization.complex_components.alert_region import AlertRegion
 
 from pyfemtet.opt._femopt_core import History
 
@@ -118,7 +118,7 @@ class HomePage(AbstractPage):
         # setup callback of subpages
         super().setup_callback()
 
-        from pyfemtet.opt.visualization2.process_monitor.application import ProcessMonitorApplication
+        from pyfemtet.opt.visualization.process_monitor.application import ProcessMonitorApplication
         from pyfemtet.opt._femopt_core import OptimizationStatus
         self.application: ProcessMonitorApplication = self.application
 
@@ -238,7 +238,7 @@ class HomePage(AbstractPage):
 class WorkerPage(AbstractPage):
 
     def __init__(self, title, rel_url, application):
-        from pyfemtet.opt.visualization2.process_monitor.application import ProcessMonitorApplication
+        from pyfemtet.opt.visualization.process_monitor.application import ProcessMonitorApplication
         self.application: ProcessMonitorApplication = None
         super().__init__(title, rel_url, application)
