+++ pyfemtet/opt/visualization/result_viewer/.gitignore
@@ -1,0 +1,1 @@
+sample
