+++ pyfemtet/opt/visualization/result_viewer/pages.py
@@ -8,11 +8,11 @@ import pandas as pd
 from dash import Output, Input, State, callback_context, no_update
 from dash.exceptions import PreventUpdate
 
-from pyfemtet.opt.visualization2.wrapped_components import dcc, dbc, html
-from pyfemtet.opt.visualization2.base import AbstractPage, logger
-from pyfemtet.opt.visualization2.complex_components.main_graph import MainGraph  # , FLEXBOX_STYLE_ALLOW_VERTICAL_FILL
-from pyfemtet.opt.visualization2.complex_components.control_femtet import FemtetControl, FemtetState
-from pyfemtet.opt.visualization2.complex_components.alert_region import AlertRegion
+from pyfemtet.opt.visualization.wrapped_components import dcc, dbc, html
+from pyfemtet.opt.visualization.base import AbstractPage, logger
+from pyfemtet.opt.visualization.complex_components.main_graph import MainGraph  # , FLEXBOX_STYLE_ALLOW_VERTICAL_FILL
+from pyfemtet.opt.visualization.complex_components.control_femtet import FemtetControl, FemtetState
+from pyfemtet.opt.visualization.complex_components.alert_region import AlertRegion
 
 from pyfemtet.opt._femopt_core import History
